GoD_FastApi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
fastapi==0.115.12
idna==3.10
pydantic==2.11.4
pydantic_core==2.33.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn[standard]


# === 추가 라이브러리 ===
aiofiles
pandas
openpyxl
numpy
httpx
tiktoken
python-dotenv
beautifulsoup4
pydantic-settings
python-multipart

